+++ nerdlets/details/index.js
@@ -4,11 +4,9 @@ import {
   PlatformStateContext,
   NerdletStateContext,
   EntityByGuidQuery,
-  HeadingText,
-  BlockText,
   Spinner
 } from 'nr1';
-import { NerdGraphError } from '@newrelic/nr1-community';
+import { NerdGraphError, EmptyState } from '@newrelic/nr1-community';
 import Breakdown from '../shared/components/breakdown';
 
 export default class Wrapper extends React.PureComponent {
@@ -46,7 +44,7 @@ export default class Wrapper extends React.PureComponent {
                     }
 
                     if (error) {
-                      return <NerdGraphError error={error} />
+                      return <NerdGraphError error={error} />;
                     }
 
                     if (
@@ -64,16 +62,13 @@ export default class Wrapper extends React.PureComponent {
                     }
 
                     return (
-                      <div className="message">
-                        <HeadingText>
-                          Site Analyzer is not available
-                        </HeadingText>
-                        <BlockText>
-                          You have access to this entity, but Site Analyzer is
-                          not enabled for Browser entities in this account.
-                          Please see your Nerdpack Manager with concerns.
-                        </BlockText>
-                      </div>
+                      <EmptyState
+                        heading="Site Analyzer is not available"
+                        description="You have access to this entity, but Site Analyzer is not
+                        enabled for Browser entities in this account. Please see
+                        your Nerdpack Manager with concerns."
+                        buttonText=""
+                      />
                     );
                   }}
                 </EntityByGuidQuery>