+++ nerdlets/root/index.js
@@ -4,12 +4,10 @@ import {
   PlatformStateContext,
   NerdletStateContext,
   EntityByGuidQuery,
-  HeadingText,
-  BlockText,
   Spinner
 } from 'nr1';
+import { EmptyState, NerdGraphError } from '@newrelic/nr1-community';
 import gql from 'graphql-tag';
-import { NerdGraphError } from '@newrelic/nr1-community';
 
 export default class Wrapper extends React.PureComponent {
   render() {
@@ -55,19 +53,16 @@ export default class Wrapper extends React.PureComponent {
                         />
                       );
                     } else if (error) {
-                      return <NerdGraphError error={error} />
+                      return <NerdGraphError error={error} />;
                     } else {
                       return (
-                        <div className="message">
-                          <HeadingText>
-                            Site Analyzer is not available
-                          </HeadingText>
-                          <BlockText>
-                            You have access to this entity, but Site Analyzer is
-                            not enabled for Browser entities in this account.
-                            Please see your Nerdpack Manager with concerns.
-                          </BlockText>
-                        </div>
+                        <EmptyState
+                          heading="Site Analyzer is not available"
+                          description="You have access to this entity, but Site Analyzer is not
+                        enabled for Browser entities in this account. Please see
+                        your Nerdpack Manager with concerns."
+                          buttonText=""
+                        />
                       );
                     }
                   }}